moving shoebox slick artifacts to the slick package
